# lib/api-zod/src/generated/types/bulkAttendanceInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Computer Department Attendance Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { BulkAttendanceInputRecordsItem } from './bulkAttendanceInputRecordsItem';

export interface BulkAttendanceInput {
  date: string;
  records: BulkAttendanceInputRecordsItem[];
}