# lib/api-zod/src/generated/types/dailyTrendPoint.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Computer Department Attendance Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface DailyTrendPoint {
  date: string;
  present: number;
  absent: number;
  late: number;
  excused: number;
  total: number;
}